Added training for RNN, but with bug
Currently, the trainer retains the hidden state between batches and
resets it in the beginning of each epoch. As result, we need to
pass retain_graph=True when doing the backward pass. As result,
there is a CUDA Out of Memory error when fully training the model
(where batch size is 256).

The overfit test fails when not retaining the hidden state between
batches.

Don't know how to solve it yet.